Results section improvements + raw sequence distance graphs
Hopefully the format of the results is easier to add information to. Deleted some sections, moved some around, general reformatting.
Addition of graphs that show frequencies of raw sequence distances within CVX (which appears to be split into two species) AND a view of the sequence distances in CVXA and CVXB, the two resulting species after the split.